Test OpenGL/Juce sous Windows : OK ! Bien meilleurs perfs.... Mais gros problème de "rétina" à résoudre (sous mac aussi) + correction bug lors de changement de scène alors qu'un déplacement touch était en cours

+++ MiamEdit/Source/GraphicSessionManager.cpp
@@ -76,7 +76,7 @@ void GraphicSessionManager::__LoadDefaultTest()
 {
     srand(2016); // GRAINE fixée ici
     
-    for(int i=0 ; i<canvasManagers.size(); i++)
+    for(size_t i=0 ; i<canvasManagers.size(); i++)
         canvasManagers[i]->__AddTestAreas();
 }
